ui: Add timeline::error module
